tests/integration/requirements.txt
cosl == 0.1.0
git+https://github.com/canonical/saml-test-idp.git
jubilant == 1.1.0
juju == 3.6.1.2
minio == 7.2.14
pytest-operator >= 0.32.0
tenacity == 9.0.0
